Update package and rename default topic
